requirements.txt
flask
gunicorn
alpaca-py

pandas
numpy

celery
redis
flower

lightgbm
prometheus-client
psutil

joblib
python-dotenv
requests
pytest
pytz